Refactor Alphabet filter UI and functionality
